Finale!

+++ Lab1_S2/Lab1_S2/src/main/java/com/aircompany/servlets/BrigadeServlet.java
@@ -5,6 +5,7 @@ import com.aircompany.db.entity.Brigade;
 import com.aircompany.db.entity.Entity;
 import com.aircompany.parsers.JsonParser;
 import com.aircompany.servlets.util.RequestPack;
+import com.aircompany.servlets.util.RoleUtil;
 
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -30,6 +31,10 @@ public class BrigadeServlet extends HttpServlet {
     public void doPut(HttpServletRequest req, HttpServletResponse resp) throws IOException{
         BufferedReader reader = req.getReader();
         String requestBodyString = RequestPack.processRequest(reader);
+        if(!RoleUtil.validateAccess(RoleUtil.getRole(req), RoleUtil.getAllowedRoles(new String[]{RoleUtil.ADMIN, RoleUtil.DISPATCH}))){
+            resp.getWriter().println("[]");
+            return;
+        }
         Entity entity = getEntity(requestBodyString);
         DaoManager DBM = new DaoManager();
         Connection conn = DBM.getConnection();
@@ -47,6 +52,10 @@ public class BrigadeServlet extends HttpServlet {
     public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException{
         BufferedReader reader = req.getReader();
         String requestBodyString = RequestPack.processRequest(reader);
+        if(!RoleUtil.validateAccess(RoleUtil.getRole(req), RoleUtil.getAllowedRoles(new String[]{RoleUtil.ADMIN, RoleUtil.DISPATCH}))){
+            resp.getWriter().println("[]");
+            return;
+        }
         Entity entity = getEntity(requestBodyString);
         if(entity == null){
             resp.getWriter().println("[]");
@@ -75,6 +84,7 @@ public class BrigadeServlet extends HttpServlet {
             resp.getWriter().println("[]");
             return;
         }
+        String role = RoleUtil.getRole(req);
         DaoManager mgr = new DaoManager();
         Connection conn = mgr.getConnection();
         if(conn == null){
@@ -84,23 +94,33 @@ public class BrigadeServlet extends HttpServlet {
         BrigadeDao brigadeDao = new BrigadeDao(conn);
         List<Entity> entityList = new ArrayList<>();
         try {
-            switch (field) {
-                case "name":
-                    entityList = brigadeDao.readByName(value);
-                    break;
-                case "id":
-                    entityList.add(brigadeDao.read(value));
-                    break;
-                case "all":
-                    entityList = brigadeDao.readAll();
-                    break;
-                case "delete":
-                    brigadeDao.delete(value);
-                    entityList = brigadeDao.readAll();
-                    break;
-                default:
-                    resp.getWriter().println("[]");
-                    return;
+            if(RoleUtil.validateAccess(role, RoleUtil.getAllowedRoles(new String[]{RoleUtil.USER, RoleUtil.DISPATCH, RoleUtil.ADMIN}))) {
+                switch (field) {
+                    case "name":
+                        entityList = brigadeDao.readByName(value);
+                        break;
+                    case "id":
+                        entityList.add(brigadeDao.read(value));
+                        break;
+                    case "all":
+                        entityList = brigadeDao.readAll();
+                        break;
+                    default:
+                        if (RoleUtil.validateAccess(role, RoleUtil.getAllowedRoles(new String[]{RoleUtil.DISPATCH, RoleUtil.ADMIN}))) {
+                            switch (field) {
+                                case "delete":
+                                    brigadeDao.delete(value);
+                                    entityList = brigadeDao.readAll();
+                                    break;
+                                default:
+                                    resp.getWriter().println("[]");
+                                    return;
+                            }
+                        } else {
+                            resp.getWriter().println("[]");
+                            return;
+                        }
+                }
             }
         }catch (Exception e){
             resp.getWriter().println("[]");
